requirements.txt
aniso8601==9.0.1
attrs==23.2.0
blinker==1.7.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
Flask==3.0.3
Flask-RESTful==0.3.10
flask-restplus==0.13.0
idna==3.7
importlib_metadata==7.1.0
itsdangerous==2.1.2
Jinja2==3.1.3
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
MarkupSafe==2.1.5
packaging==24.0
pillow==10.3.0
pip-check==2.8.1
pytz==2024.1
referencing==0.34.0
requests==2.31.0
rpds-py==0.18.0
six==1.16.0
terminaltables==3.1.10
urllib3==2.2.1
Werkzeug==3.0.2
zipp==3.18.1
